feat: auto restore hovered folder
Now the hovered folder is restore if you enter/leave a folder with
hidden = true to a folder with hidden = false. Example: from
`/home/a/.config/` to `/home/a/`. The hovered folder `.config`.  Or
from `/home/a/` (hovered over `.config` folder) to `/home` then go back
to `/home/a/` the hovered folder still `.config`.